Factor out runner into its own package in preparation for forking daemon mode into its own main().
